+++ CHANGES.md
@@ -1,3 +1,12 @@
+# v1.2.3
+* Add more multi-fidelity plots
+
+# v1.2.2
+* Revising plotting of multifidelity work
+
+# v1.2.1
+* Release for minting DOI
+
 # v1.2.0
 * Bug fix